+++ src/main/java/controller/CreateUserController.java
@@ -1,4 +1,4 @@
-package webserver.controller;
+package controller;
 
 import db.DataBase;
 import exception.RequestParameterNotFoundException;
